main 삭제, menuform, menulist로 컴포넌트 분리
